requirements.txt
asgiref==3.3.4
Django==3.2
numpy==1.20.2
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
